+++ backend/src/main/java/fk/prof/backend/aggregator/CpuSamplingAggregationBucket.java
@@ -3,7 +3,7 @@ package fk.prof.backend.aggregator;
 import fk.prof.aggregation.FinalizableBuilder;
 import fk.prof.aggregation.finalized.FinalizedCpuSamplingAggregationBucket;
 import fk.prof.backend.exception.AggregationFailure;
-import fk.prof.backend.model.request.RecordedProfileIndexes;
+import fk.prof.backend.model.profile.RecordedProfileIndexes;
 import fk.prof.aggregation.MethodIdLookup;
 import fk.prof.aggregation.cpusampling.CpuSamplingFrameNode;
 import fk.prof.aggregation.cpusampling.CpuSamplingTraceDetail;
