+++ backend/src/main/java/fk/prof/backend/model/policy/PolicyDetails.java
@@ -1,4 +1,4 @@
-package fk.prof;
+package fk.prof.backend.model.policy;
 
 import java.time.LocalDateTime;
