+++ backend/src/main/java/fk/prof/backend/model/profile/RecordedProfileHeader.java
@@ -1,4 +1,4 @@
-package fk.prof.backend.model.request;
+package fk.prof.backend.model.profile;
 
 import recording.Recorder;
