+++ backend/src/main/java/fk/prof/backend/request/profile/RecordedProfileRequestHandler.java
@@ -1,6 +1,7 @@
-package fk.prof.backend.model.request;
+package fk.prof.backend.request.profile;
 
 import fk.prof.backend.exception.HttpFailure;
+import fk.prof.backend.request.CompositeByteBufInputStream;
 import fk.prof.backend.verticles.http.HttpHelper;
 import io.vertx.core.Handler;
 import io.vertx.core.buffer.Buffer;
@@ -30,10 +31,6 @@ public class RecordedProfileRequestHandler implements Handler<Buffer> {
       inputStream.accept(requestBuffer.getByteBuf());
       try {
         profileParser.process(inputStream);
-        //NOTE: Do not rely on CodedInputStream::getTotalBytesRead() to determine remaining bytes to be parsed
-        //Example: CodedInputStream::readUInt32 can throw InvalidProtocolBufferEx if incomplete bytes have been received but it will update the totalBytesRead count
-        //This will result in un-parsed bytes getting discarded and not accounted for when next chunk is received
-        //Maintaining own counter "currentPos" which serves as a checkpoint = last successfully read byte in the buffer + 1
       } catch (Exception ex) {
         try {
           inputStream.close();
